pyproject.toml
[project]
name = "pydwm1001"
version = "0.12.1"
authors = [
  { name="Adam Morrissett", email="morrissettal2@vcu.edu" },
]
description = "A Python library for interfacing with DWM1001"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "pyserial"
]

[project.urls]
"Homepage" = "https://github.com/the-hive-lab/pydwm1001"
"Bug Tracker" = "https://github.com/the-hive-lab/pydwm1001/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
